Update text_to_speech_python_code
Welcome to my Python text-to-speech repository! This project includes code for converting written text into spoken audio using Python and the Google Text-to-Speech API. With just a few lines of code, you can easily integrate text-to-speech functionality into your project. Whether you're building a chatbot or creating an automated voice system, this code will save you time and effort in implementing this feature. Simply input the text you want to be spoken and let the API do the rest!